Added two new API functions for clixon and modifed build for autotools
This reason for this is to run the evhtp parser WITHOUT its SSL usage and libevent
event loop. In this way, Clixon calls SSL itself and uses its internal event loop.

The two new functions are:
- int connection_parse_nobev(char *buf, size_t buflen, void *arg);
- evhtp_connection_t *evhtp_connection_new_server(evhtp_t * htp, evutil_socket_t sock);

The reason for using autotools instead of cmake is convenience only since autotools is standard
in the clixon world.